commit2

+++ Ado.net homework2/Ado.net homework2/Program.cs
@@ -31,7 +31,11 @@ do
                     Console.WriteLine("FullName :");
                     string fullname = Console.ReadLine();
 
-
+                    if (string.IsNullOrWhiteSpace(fullname))
+                    {
+                        Console.WriteLine("Bos ola bilmez");
+                        break;
+                    }
                     Console.WriteLine("Position :");
                     string position = Console.ReadLine();
 
@@ -54,7 +58,13 @@ do
                 case "b":
 
                     Console.WriteLine("Id :");
-                    int id = Convert.ToInt32(Console.ReadLine());
+                    string idstro = Console.ReadLine();
+                    int id;
+                    if (!int.TryParse(idstro,out id))
+                    {
+                        Console.WriteLine("Duzgun daxil edin");
+                        break;
+                    }
 
                     Console.WriteLine("FullName :");
                     fullname = Console.ReadLine();
@@ -81,7 +91,12 @@ do
                     break;
                 case "c":
                     Console.WriteLine("Id :");
-                    id = Convert.ToInt32(Console.ReadLine());
+                    string idopt2 = Console.ReadLine();
+                    if (!int.TryParse(idopt2,out id))
+                    {
+                        Console.WriteLine("Duzgun id daxil edin");
+                        break;
+                    }
 
                     var data = speakersDao.GetSpeakersById(id);
                     if (data == null) Console.WriteLine("Speaker tapilmadi");
@@ -96,7 +111,12 @@ do
                     break;
                 case "e":
                     Console.WriteLine("Id :");
-                    id = Convert.ToInt32(Console.ReadLine());
+                    string idopt3 = Console.ReadLine();
+                    if (!int.TryParse(idopt3,out id))
+                    {
+                        Console.WriteLine("Duzgun id daxil edin");
+                        break;
+                    }
                     speakersDao.DeleteSpeakersById(id);
                     break;
                 default:
@@ -108,6 +128,7 @@ do
             Console.WriteLine("a : Insert Events");
             Console.WriteLine("b : Get By id");
             Console.WriteLine("c : GettAll");
+            Console.WriteLine("d : Add EventSpeaker");
             string opt3 = Console.ReadLine();
             switch (opt3)
             {
@@ -115,6 +136,11 @@ do
                     Console.WriteLine("Name :");
                     string nameOpt = Console.ReadLine();
 
+                    if (string.IsNullOrWhiteSpace(nameOpt))
+                    {
+                        Console.WriteLine("Bosa ola bilmez");
+                        break;
+                    }
 
                     Console.WriteLine("Description :");
                     string descOpt = Console.ReadLine();
@@ -123,16 +149,33 @@ do
                     string addOpt = Console.ReadLine();
 
                     Console.WriteLine("StartDate :");
-                    DateTime startdate = Convert.ToDateTime(Console.ReadLine());
+                    string startdateopt = Console.ReadLine();
+                    DateTime startdate;
+
+                    if (!DateTime.TryParse(startdateopt,out startdate))
+                    {
+                        Console.WriteLine("Duzgun date daxil edin");
+                        break;
+                    }
+                    
 
                     Console.WriteLine("StartTime :");
-                    DateTime starttime = Convert.ToDateTime(Console.ReadLine());
+                    string starttimeopt = Console.ReadLine();
+                    DateTime starttime;
+                    if (!DateTime.TryParse(starttimeopt,out starttime))
+                    {
+                        Console.WriteLine("Duzgun date daxil edin");
+                        break;
+                    }
 
                     Console.WriteLine("EndTime");
                     string endtimeOpt = Console.ReadLine();
-                    DateTime endtime = Convert.ToDateTime(endtimeOpt);
-
-
+                    DateTime endtime;
+                    if (!DateTime.TryParse(endtimeOpt,out endtime))
+                    {
+                        Console.WriteLine("Duzgun date daxil edin");
+                        break;
+                    }
 
                     Events events = new Events()
                     {
@@ -147,7 +190,13 @@ do
                     break;
                 case "b":
                     Console.WriteLine("Id :");
-                    int id = Convert.ToInt32(Console.ReadLine());
+                    string idstr = Console.ReadLine();
+                    int id;
+                    if (!int.TryParse(idstr,out id))
+                    {
+                        Console.WriteLine("Duzgun daxil edin");
+                        break;
+                    }
 
                     var data = eventDao.GetEventsById(id);
                     if (data == null) Console.WriteLine("Evnt tapilmadi");
@@ -158,6 +207,27 @@ do
                     {
                         Console.WriteLine(item);
                     }
+                    break;
+                case "d":
+                    Console.WriteLine("Speaker Id:");
+                    string speakeridOpt = Console.ReadLine();
+                    int speakerid;
+                    if (!int.TryParse(speakeridOpt,out speakerid))
+                    {
+                        Console.WriteLine("Duzgun Id daxil edin");
+                        break;
+                    }
+
+                    Console.WriteLine("Event Id:");
+                    string eventidopt = Console.ReadLine();
+                    int eventid;
+                    if (!int.TryParse(eventidopt,out eventid))
+                    {
+                        Console.WriteLine("Duzgun daxil edin");
+                        break;
+                    }
+
+                    eventDao.AddSpeakerAndEvent(speakerid, eventid);
                     break;
                 default:
                     Console.WriteLine("\nDuzgun opt secin\n");
